Split out envvar into separate project_envvar file
